requirements.txt
bidict==0.22.0
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
eventlet==0.33.2
Flask==2.2.2
Flask-SocketIO==5.3.2
greenlet==3.3.0
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpickle==3.0.0
MarkupSafe==2.1.1
python-engineio==4.3.4
python-socketio==5.7.2
simple-websocket==1.0.0
six==1.16.0
Werkzeug==2.2.2
zipp==3.11.0
gunicorn==21.2.0
PyGithub==2.1.1
supabase==2.27.0
python-dotenv==1.0.0
websockets==15.0.1
